+++ ql/cashflows/overnightindexedcouponpricer.hpp
@@ -4,6 +4,7 @@
  Copyright (C) 2009 Roland Lichters
  Copyright (C) 2009 Ferdinando Ametrano
  Copyright (C) 2014 Peter Caspers
+ Copyright (C) 2016 Stefano Fondi
  Copyright (C) 2017 Joseph Jeisman
  Copyright (C) 2017 Fabrice Lecuyer
 
@@ -51,6 +52,39 @@ namespace QuantLib {
 
       protected:
         const OvernightIndexedCoupon* coupon_ = nullptr;
+    };
+
+    /*! pricer for arithmetically averaged overnight indexed coupons
+    Reference: Katsumi Takada 2011, Valuation of Arithmetically Average of
+    Fed Funds Rates and Construction of the US Dollar Swap Yield Curve
+    */
+    class ArithmeticAveragedOvernightIndexedCouponPricer : public FloatingRateCouponPricer {
+      public:
+        explicit ArithmeticAveragedOvernightIndexedCouponPricer(
+            Real meanReversion = 0.03,
+            Real volatility = 0.00, // NO convexity adjustment by default
+            bool byApprox = false)  // TRUE to use Katsumi Takada approximation
+        : byApprox_(byApprox), mrs_(meanReversion), vol_(volatility) {}
+
+        explicit ArithmeticAveragedOvernightIndexedCouponPricer(
+            bool byApprox) // Simplified constructor assuming no convexity correction
+        : ArithmeticAveragedOvernightIndexedCouponPricer(0.03, 0.0, byApprox) {}
+
+        void initialize(const FloatingRateCoupon& coupon) override;
+        Rate swapletRate() const override;
+        Real swapletPrice() const override { QL_FAIL("swapletPrice not available"); }
+        Real capletPrice(Rate) const override { QL_FAIL("capletPrice not available"); }
+        Rate capletRate(Rate) const override { QL_FAIL("capletRate not available"); }
+        Real floorletPrice(Rate) const override { QL_FAIL("floorletPrice not available"); }
+        Rate floorletRate(Rate) const override { QL_FAIL("floorletRate not available"); }
+
+      protected:
+        Real convAdj1(Time ts, Time te) const;
+        Real convAdj2(Time ts, Time te) const;
+        const OvernightIndexedCoupon* coupon_;
+        bool byApprox_;
+        Real mrs_;
+        Real vol_;
     };
 }
